feat(MediaQuery): add Typescript types for consts
This is needed to correctly type the Queries object (using it for SSR)
