+++ src/utils/mediaQuery/consts.d.ts
@@ -1,0 +1,17 @@
+declare module "@kiwicom/orbit-components/lib/utils/mediaQuery/consts";
+
+export type Devices =
+  | "largeDesktop"
+  | "desktop"
+  | "tablet"
+  | "largeMobile"
+  | "mediumMobile"
+  | "smallMobile";
+
+declare const QUERIES: {
+  LARGEDESKTOP: "largeDesktop";
+  DESKTOP: "desktop";
+  TABLET: "tablet";
+  LARGEMOBILE: "largeMobile";
+  MEDIUMMOBILE: "mediumMobile";
+};
